Modify Project Workflows (#8)
* Updated Trigger Conditions for Code_Coverage

* Updated Runner information for all Workflos

* Refactored RocProfiler-SDK CodeQL

* Added Sparse Checkout

* Grab submodules from rocprofiler-sdk

* Update Submodules

* Moved gitmodules to root

* Update the submodule path of fmt

* Remove fetch for submodules

* Added fmt to gitmodules

* Added glog to submodules

* Added remaining submodules for rocprofiler-sdk

* Added Filesystem to Submodules

* Added remaining submodules

* Update Requirements File

* Updated docs.yml

* Added Workflows to Sparse Checkout

* Fixed Sparse Checkout

* Removed Workflows from checkout

* Checkout specific actions files

* Refactored docs.yml

* Update rocprofiler-sdk-docs.yml

* Removed Unnecessary Workflows

* Added RocProfiler-SDK Restrictions

* Added rocprofiler-sdk Code Coverage

* Added rocprofiler-sdk CI Workflow

* Update rocprofiler-sdk-continuous_integration.yml

* Update rocprofiler-sdk-continuous_integration.yml

Added path

* Added the pipe operator to fix breaks

* Updated Working Directories

* Added KWS Caller Workflow

* Added AqlProfile CodeQL

* Added permissions to the KWS

* Added ROCm-CI Caller

* Added RocProfiler Rocm CI Caller

* Added Python Linting Workflows

* Added ROCm Release Compatibility Workflow

* Fix Typo For Python Linting

* Added Path to trigger Python Linting for Rocprofiler-Systems

* Added rocprofiler-systems Formatting

* Added Formatting to Main Workflows

* Removed Python 3.7 from rocprofiler-systems

* Updated ROCm release compatibility workflow

* Fixed path in Install Step

* Change Safe Dir in CI Workflow

* Added CI Workflow for rocprofiler-register

* Added Containers Workflow

* Fix compute docs

* Changing Runners

* Added Workflows to Project Subfolders

* Update rocprofiler-sdk-continuous_integration.yml

---------

+++ projects/aqlprofile/gfxip/gfx12/gfx12_primitives.h
@@ -560,7 +560,7 @@ class gfx12_cntx_prim {
 
   // Indicates the size of buffer to use per Shader Engine instance.
   // The size is specified in terms of 4KB blocks
-  static uint32_t sqtt_buffer0_size_value(uint64_t size_val) {
+  static uint32_t sqtt_buffer0_size_value(uint32_t size_val) {
     uint32_t sq_thread_trace_buf0_size{0};
     sq_thread_trace_buf0_size =
         SET_REG_FIELD_BITS(SQ_THREAD_TRACE_BUF0_SIZE, SIZE, size_val >> TT_BUFF_ALIGN_SHIFT);
